refactor(storage): extract adjacency spill builder (#1181)
